Add type annotations to let expressions
