+++ inferno-core/src/Inferno/Eval.hs
@@ -169,6 +169,11 @@ eval env@(localEnv, pinnedEnv) expr = case expr of
         argv <- eval env arg
         f argv
       _ -> throwM $ RuntimeError "failed to match with a function"
+  -- TODO create pattern for LetAnnot, or reuse?
+  LetAnnot _ _ x _ _ _ e _ body -> do
+    e' <- eval env e
+    let nenv = Map.insert x e' localEnv
+    eval (nenv, pinnedEnv) body
   Let_ (Expl x) e body -> do
     e' <- eval env e
     let nenv = Map.insert x e' localEnv
